requirements.txt
python-dotenv
deep-translator
pandas
streamlit
langchain-groq
pandasai
openpyxl